remove old shields files
