+++ components/brave_extension/extension/brave_extension/BUILD.gn
@@ -41,18 +41,10 @@ transpile_web_ui("brave_extension") {
     "background/store.ts",
     "braveShieldsPanel.tsx",
     "components/braveShields.tsx",
-    "components/braveShields2.tsx",
     "components/footer.tsx",
-    "components/footer2.tsx",
     "components/header.tsx",
-    "components/header2.tsx",
     "components/interfaceControls.tsx",
-    "components/interfaceControls2.tsx",
     "components/privacyControls.tsx",
-    "components/privacyControls2.tsx",
-    "components/blockedResources/blockedResources.tsx",
-    "components/blockedResources/scriptsList.tsx",
-    "components/blockedResources/staticList.tsx",
     "components/controls/adsTrackersControl.tsx",
     "components/controls/cookiesControl.tsx",
     "components/controls/deviceRecognitionControl.tsx",
