+++ components/brave_extension/extension/brave_extension/containers/braveShields.tsx
@@ -6,7 +6,7 @@ import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
 import * as shieldsPanelActions from '../actions/shieldsPanelActions'
 import * as shieldsPanelState from '../state/shieldsPanelState'
-import BraveShields from '../components/braveShields2'
+import BraveShields from '../components/braveShields'
 import { State } from '../types/state/mainState'
 
 const mapStateToProps = (state: State) => ({
